Capture file path from test definition callsite
Prior to this commit the code was capturing the test file path when `test.run()`
was called, but this is not necessarily going to be the same file as where the
tests are defined. This change makes it so the test file path is captured where
the test itself is defined.